feat(plugins): remove config and allow plugin elements (#12)
* feat(plugins): remove config and allow plugin elements

Plugins can bve nested in a special component. They are filled with additional props and gain access
to the context (including the store)

* feat(plugins): give access to public store methods

plugins should not have to access the context as this could change in future versions
Listener was ported to the new plugin mechanism

* docs(plugins): gh-pages example uses Plugins

Show how plugins work and can be used in the gh-pages example app.